Rename C source and header files
For consistency with GTK, blah.[ch] should be named eosblah.[ch].

[#23]
