+++ docs/reference/endless/Makefile.am
@@ -48,7 +48,7 @@ EXTRA_HFILES=
 
 # Header files or dirs to ignore when scanning. Use base file/dir names
 # e.g. IGNORE_HFILES=gtkdebug.h gtkintl.h private_code
-IGNORE_HFILES=$(top_srcdir)/endless/init-private.h
+IGNORE_HFILES=$(top_srcdir)/endless/eosinit-private.h
 
 # Images to copy into HTML directory.
 # e.g. HTML_IMAGES=$(top_srcdir)/gtk/stock-icons/stock_about_24.png
