+++ endless/eosenums.h
