+++ endless/eosinit-private.h
