+++ endless/eosmacros.h
